--- keewe-api/src/main/java/ccc/keeweapi/api/insight/ReactionController.java
@@ -1,23 +1,0 @@
-package ccc.keeweapi.api.insight;
-
-import ccc.keeweapi.dto.ApiResponse;
-import ccc.keeweapi.dto.insight.ReactRequest;
-import ccc.keeweapi.dto.insight.ReactResponse;
-import ccc.keeweapi.service.insight.ReactionApiService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RequiredArgsConstructor
-@RequestMapping("/api/v1/reaction")
-@RestController
-public class ReactionController {
-    private final ReactionApiService reactionApiService;
-
-    @PostMapping
-    public ApiResponse<ReactResponse> react(@RequestBody ReactRequest request) {
-        return ApiResponse.ok(reactionApiService.react(request));
-    }
-}
