+++ keewe-core/src/main/java/ccc/keewecore/utils/ObjectMapperUtils.java
@@ -2,12 +2,14 @@ package ccc.keewecore.utils;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import java.io.IOException;
 
 @Component
+@Slf4j
 public class ObjectMapperUtils {
     private static ObjectMapper objectMapper;
 
@@ -20,14 +22,25 @@ public class ObjectMapperUtils {
         try {
             return objectMapper.writeValueAsString(src);
         } catch (JsonProcessingException ex) {
+            log.error(ex.getMessage());
             return "";
+        }
+    }
+
+    public static <T> byte[] writeValueAsBytes(T src) {
+        try {
+            return objectMapper.writeValueAsBytes(src);
+        } catch (JsonProcessingException ex) {
+            log.error(ex.getMessage(), ex);
+            return null;
         }
     }
 
     public static <T> T readValue(String src, Class<T> clazz) {
         try {
             return objectMapper.readValue(src, clazz);
-        } catch (IOException e) {
+        } catch (IOException ex) {
+            log.error(ex.getMessage(), ex);
             return null;
         }
     }
@@ -35,7 +48,8 @@ public class ObjectMapperUtils {
     public static <T> T readValue(byte[] src, Class<T> clazz) {
         try {
             return objectMapper.readValue(src, clazz);
-        } catch (IOException e) {
+        } catch (IOException ex) {
+            log.error(ex.getMessage(), ex);
             return null;
         }
     }
