+++ keewe-domain/src/main/java/ccc/keewedomain/service/title/AbstractTitleStatService.java
@@ -4,12 +4,15 @@ import ccc.keewecore.consts.KeeweConsts;
 import ccc.keewecore.consts.TitleCategory;
 import ccc.keewecore.dto.TitleEvent;
 import ccc.keewecore.utils.KeeweTitleHeader;
+import ccc.keewecore.utils.ObjectMapperUtils;
 import ccc.keewedomain.persistence.domain.title.Title;
 import ccc.keewedomain.persistence.domain.title.TitleAchievement;
 import ccc.keewedomain.persistence.domain.user.User;
 import ccc.keewedomain.persistence.repository.user.TitleAchievementRepository;
 import ccc.keewedomain.service.user.UserDomainService;
 import ccc.keeweinfra.service.MQPublishService;
+import org.springframework.amqp.core.Message;
+import org.springframework.amqp.core.MessageBuilder;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
@@ -33,20 +36,21 @@ public abstract class AbstractTitleStatService<T, S> {
     public void aggregateStat(KeeweTitleHeader header) {
         T result = aggregate(header);
 
-        getTitleOpts(result).ifPresent(s -> {
+        getAcquiredTitleOps(result).ifPresent(s -> {
                     Title title = saveAndGetTitle(s);
 
                     User user = userDomainService.getUserByIdOrElseThrow(Long.valueOf(header.getUserId()));
                     titleAchievementRepository.save(TitleAchievement.of(user, title));
 
                     TitleEvent event = TitleEvent.of(title.getCategory(), title.getName(),title.getIntroduction(), LocalDateTime.now());
-                    mqPublishService.publish(KeeweConsts.TITLE_ACQUIREMENT_EXCHANGE, event);
+                    Message message = MessageBuilder.withBody(ObjectMapperUtils.writeValueAsBytes(event)).build();
+                    mqPublishService.publish(KeeweConsts.TITLE_ACQUIREMENT_EXCHANGE, KeeweConsts.DEFAULT_ROUTING_KEY, message, header::toMessageWithHeader);
                 });
     }
 
     protected abstract T aggregate(KeeweTitleHeader header);
 
-    protected abstract Optional<S> getTitleOpts(T target);
+    protected abstract Optional<S> getAcquiredTitleOps(T target);
 
     protected abstract Title saveAndGetTitle(S param);
