+++ keewe-infra/src/main/java/ccc/keeweinfra/service/MQPublishService.java
@@ -4,6 +4,7 @@ import ccc.keewecore.consts.KeeweConsts;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.amqp.core.Message;
+import org.springframework.amqp.core.MessagePostProcessor;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.stereotype.Service;
 
@@ -15,15 +16,11 @@ public class MQPublishService {
 
     //TODO: retry handler 도입
     public <T> void publish(String exchange, T body) {
-        rabbitTemplate.convertAndSend(exchange
-                , KeeweConsts.DEFAULT_ROUTING_KEY
-                , body);
+        rabbitTemplate.convertAndSend(exchange, KeeweConsts.DEFAULT_ROUTING_KEY, body);
     }
 
-    public <T> void publish(String queueName, Message message) {
-        rabbitTemplate.convertAndSend(
-                KeeweConsts.DEFAULT_ROUTING_KEY
-                , queueName
-                , message);
+    public void publish(String exchange, String queueName, Message message, MessagePostProcessor messagePostProcessor) {
+        rabbitTemplate.convertAndSend(exchange, queueName, message, messagePostProcessor);
     }
+
 }
