+++ keewe-push/src/main/java/ccc/keewepush/KeewePushApplication.java
@@ -2,8 +2,10 @@ package ccc.keewepush;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
+@ComponentScan("ccc")
 public class KeewePushApplication {
 
     public static void main(String[] args) {
