+++ keewe-push/src/main/resources/application.yml
@@ -2,6 +2,8 @@ spring:
   profiles:
     include: domain, core
 
+server:
+  port: 5050
 
 ---
